Massive Overhaul
Re-specialized the Kernel data. Implemented DataFile struct in utility,
but may bring that over to kernel for its specificity. Implemented
random_shuffle on the easier to randomize objects. Still need to
implement it with the fftext files however. That, fixing materia, and
maybe some work on scene will be in the next commit.