Sub-file Restructure
Merged all of the structuring methods, cutting about 200 lines and massively reducing the complexity of the code base. Doing this also centralized the declaration of how the data is structured on Kernel object initialization. In addition, those have been changed to constants as they should never need to be changed.